requirements.txt
beautifulsoup4==4.6.1
Django==2.0.7
djangorestframework==3.8.2
pytz==2018.5
